+++ src/Rask.Wasm/Browser/RaskWasmBrowserJsonContext.cs
@@ -13,4 +13,6 @@ namespace Rask.Wasm.Browser;
 [JsonSourceGenerationOptions(JsonSerializerDefaults.Web)]
 [JsonSerializable(typeof(ShareData))]
 [JsonSerializable(typeof(PushSubscription))]
+[JsonSerializable(typeof(NotificationOptions))]
+[JsonSerializable(typeof(OrientationReading))]
 internal sealed partial class RaskWasmBrowserJsonContext : JsonSerializerContext;